+++ src/test/io_test.py
@@ -1,3 +1,5 @@
+import pytest
+
 from olmo_core.io import (
     deserialize_from_tensor,
     list_directory,
@@ -21,7 +23,7 @@ def test_list_local_directory(tmp_path):
     assert set(list_directory(tmp_path)) == {f"{tmp_path}/file1.json", f"{tmp_path}/dir1"}
 
 
-def test_list_remote_directory(tmp_path, s3_checkpoint_dir):
+def _run_list_remote_directory(tmp_path, remote_dir):
     (tmp_path / "file1.json").touch()
     (tmp_path / "dir1").mkdir()
     (tmp_path / "dir1" / "file2").touch()
@@ -30,11 +32,24 @@ def test_list_remote_directory(tmp_path, s3_checkpoint_dir):
         if not path.is_file():
             continue
         rel_path = path.relative_to(tmp_path)
-        upload(path, f"{s3_checkpoint_dir}/{rel_path}")
+        upload(path, f"{remote_dir}/{rel_path}")
 
     # Should only list immediate children (files and dirs), but not files in subdirs.
     # The paths returned should be full paths.
-    assert set(list_directory(s3_checkpoint_dir)) == {
-        f"{s3_checkpoint_dir}/file1.json",
-        f"{s3_checkpoint_dir}/dir1",
+    assert set(list_directory(remote_dir)) == {
+        f"{remote_dir}/file1.json",
+        f"{remote_dir}/dir1",
     }
+
+
+def test_list_remote_directory_s3(tmp_path, s3_checkpoint_dir):
+    _run_list_remote_directory(tmp_path, s3_checkpoint_dir)
+
+
+def test_list_remote_directory_gcs(tmp_path, gcs_checkpoint_dir):
+    from google.auth.exceptions import DefaultCredentialsError
+
+    try:
+        _run_list_remote_directory(tmp_path, gcs_checkpoint_dir)
+    except DefaultCredentialsError:
+        pytest.skip("Requires authentication with Google Cloud")